# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "batsim",
    "description": "Residential battery fleet simulator: physics-faithful virtual homes behind an OpenAPI-first HTTP API for dispatch-strategy testing (ERCOT).",
    "license": {
      "name": "MIT OR Apache-2.0",
      "identifier": "MIT OR Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/v1/dispatch": {
      "post": {
        "tags": [
          "dispatch"
        ],
        "summary": "Send a command to a home/fleet subset.",
        "operationId": "dispatch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DispatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Command accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DispatchResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Unknown target",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Idempotency conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Empty target set",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/dispatch/commands": {
      "get": {
        "tags": [
          "dispatch"
        ],
        "summary": "Audit log.",
        "operationId": "list_commands",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Page size.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "maximum": 1000,
              "minimum": 1
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Continuation cursor.",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(?:[A-Za-z0-9_-]{4})*(?:[A-Za-z0-9_-]{2,3})?$"
            }
          },
          {
            "name": "target",
            "in": "query",
            "description": "Filter by target home id.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filter by rollup status.",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/CommandStatus"
            }
          },
          {
            "name": "since",
            "in": "query",
            "description": "Only commands accepted at or after this RFC 3339 time.",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Page of commands",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandsPage"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/dispatch/commands/{command_id}": {
      "get": {
        "tags": [
          "dispatch"
        ],
        "summary": "Command status and per-target execution detail.",
        "operationId": "get_command",
        "parameters": [
          {
            "name": "command_id",
            "in": "path",
            "description": "Command id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The command",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandDoc"
                }
              }
            }
          },
          "404": {
            "description": "Unknown command",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "dispatch"
        ],
        "summary": "Cancel a command's still-queued targets.",
        "operationId": "cancel_command",
        "parameters": [
          {
            "name": "command_id",
            "in": "path",
            "description": "Command id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Command cancelled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkDoc"
                }
              }
            }
          },
          "404": {
            "description": "Unknown command",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Command already finished",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CommandDoc": {
        "type": "object",
        "description": "Audit-log command record.",
        "required": [
          "command_id",
          "status",
          "created_at",
          "principal",
          "request_hash",
          "request",
          "targets"
        ],
        "properties": {
          "command_id": {
            "type": "string",
            "description": "Command id."
          },
          "created_at": {
            "type": "string",
            "description": "Wall-clock acceptance time."
          },
          "idempotency_key": {
            "type": [
              "string",
              "null"
            ],
            "description": "Idempotency key, when supplied."
          },
          "principal": {
            "type": "string",
            "description": "Requesting principal."
          },
          "request": {
            "$ref": "#/components/schemas/DispatchRequest",
            "description": "The original request."
          },
          "request_hash": {
            "type": "string",
            "description": "SHA-256 of the canonical request body."
          },
          "status": {
            "$ref": "#/components/schemas/CommandStatus",
            "description": "Rollup status."
          },
          "targets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TargetExecution"
            },
            "description": "Per-target execution detail."
          }
        }
      },
      "CommandStatus": {
        "type": "string",
        "description": "Command rollup status.",
        "enum": [
          "queued",
          "in_flight",
          "completed",
          "completed_with_errors",
          "cancelled"
        ]
      },
      "CommandsPage": {
        "type": "object",
        "description": "Page of command records.",
        "required": [
          "data",
          "page"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommandDoc"
            },
            "description": "Commands."
          },
          "page": {
            "$ref": "#/components/schemas/PageInfo",
            "description": "Pagination."
          }
        }
      },
...
      "DispatchRequest": {
        "type": "object",
        "description": "Dispatch request.",
        "required": [
          "target",
          "action"
        ],
        "properties": {
          "action": {
            "$ref": "#/components/schemas/ActionSpec",
            "description": "Action."
          },
          "command_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Client-supplied command id (ULID/uuid). Retries with the same id\nare deduplicated before enqueue. Server-assigned when omitted."
          },
          "execution": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ExecutionSpec",
                "description": "Execution shaping."
              }
            ]
          },
          "target": {
            "$ref": "#/components/schemas/TargetSpec",
            "description": "Target set."
          }
        },
        "additionalProperties": false
      },
      "DispatchResponse": {
        "type": "object",
        "description": "Dispatch acceptance response.",
        "required": [
          "command_id",
          "accepted",
          "targets",
          "status",
          "status_url"
        ],
        "properties": {
          "accepted": {
            "type": "boolean",
            "description": "Always true for a 202 response."
          },
          "command_id": {
            "type": "string",
            "description": "The command id (echoed or server-assigned)."
          },
          "status": {
            "$ref": "#/components/schemas/CommandStatus",
            "description": "Rollup status at response time."
          },
          "status_url": {
            "type": "string",
            "description": "Poll URL for execution detail."
          },
          "targets": {
            "type": "integer",
            "description": "Number of homes targeted.",
            "minimum": 0
          }
        }
      },
      "ExecutionSpec": {
        "type": "object",
        "description": "Execution shaping for a command.",
        "properties": {
          "latency_ms": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/LatencySpec",
                "description": "Per-device execution latency; each home draws its own delay,\nmimicking a vendor cloud API. Default uniform 250..2000 ms."
              }
            ]
          },
          "ramp": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/RampStyle",
                "description": "Ramp style; only `immediate` is supported."
              }
            ]
          },
          "timeout_s": {
            "type": [
              "integer",
              "null"
            ],
            "description": "Per-device timeout in seconds (default 30).",
            "maximum": 18446744073709551615,
            "minimum": 0
          }
        },
        "additionalProperties": false
      },
...
      "OkDoc": {
        "type": "object",
        "description": "Empty success body for action endpoints with nothing to return.",
        "required": [
          "ok"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "description": "Always true."
          }
        }
      },
...
      "PageInfo": {
        "type": "object",
        "description": "Cursor pagination envelope.",
        "required": [
          "has_more"
        ],
        "properties": {
          "has_more": {
            "type": "boolean",
            "description": "Whether more results exist."
          },
          "next_cursor": {
            "type": [
              "string",
              "null"
            ],
            "description": "Opaque cursor for the next page; null when exhausted."
          }
        }
      },
...
      "Problem": {
        "type": "object",
        "description": "RFC 9457 problem document; the only error body the API emits.",
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ProblemCode",
            "description": "Stable machine-readable code."
          },
          "detail": {
            "type": [
              "string",
              "null"
            ],
            "description": "Specifics of this occurrence."
          },
          "instance": {
            "type": [
              "string",
              "null"
            ],
            "description": "The request path that produced the problem."
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "HTTP status code.",
            "minimum": 0
          },
          "title": {
            "type": "string",
            "description": "Short human-readable summary."
          },
          "trace_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Request correlation id (echoed as `X-Request-Id`)."
          },
          "type": {
            "type": "string",
            "description": "Problem type URN."
          }
        },
        "additionalProperties": true
      },
      "ProblemCode": {
        "type": "string",
        "description": "Stable machine-readable error codes.",
        "enum": [
          "VALIDATION_ERROR",
          "UNAUTHORIZED",
          "NOT_FOUND",
          "CONFLICT",
          "UNPROCESSABLE",
          "SIM_RUNNING",
          "SIM_NOT_RUNNING",
          "IDEMPOTENCY_KEY_REUSE",
          "INTERNAL"
        ]
      },
...
      "RampStyle": {
        "type": "string",
        "description": "Ramp styles for power commands.",
        "enum": [
          "immediate"
        ]
      },
...
      "TargetSpec": {
        "anyOf": [
          {
            "type": "object",
            "required": [
              "fleet_id"
            ],
            "properties": {
              "filter": {
                "type": [
                  "object",
                  "null"
                ]
              },
              "fleet_id": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "home_ids": {
                "type": [
                  "array",
                  "null"
                ],
                "items": {
                  "type": "string"
                }
              },
              "sample_pct": {
                "type": [
                  "number",
                  "null"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "home_ids"
            ],
            "properties": {
              "filter": {
                "type": [
                  "object",
                  "null"
                ]
              },
              "fleet_id": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "home_ids": {
                "type": [
                  "array",
                  "null"
                ],
                "items": {
                  "type": "string"
                }
              },
              "sample_pct": {
                "type": [
                  "number",
                  "null"
                ]
              }
            }
          }
        ],
        "description": "Dispatch target set: a fleet id, explicit home ids, or both."
      },
...
    }
  },
...
}